third_party/WebKit/LayoutTests/bluetooth/characteristic/readValue/blocklisted-characteristic.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../external/wpt/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Characteristic with exclude-reads fullfills write and ' +
    'rejects read.';
const expected = new DOMException('readValue() called on blocklisted object ' +
    'marked exclude-reads. https://goo.gl/4NeimX', 'SecurityError');
let characteristic;

bluetooth_test(() => getBlocklistExcludeReadsCharacteristic()
    // 1. Connect to the device and retrieve the GATT characteristic
    .then(_ => ({characteristic} = _))
    // 2. Attempt to write and read the characteristic
    .then(() => characteristic.writeValue(new Uint8Array(1 /* length */)))
    .then(() => assert_promise_rejects_with_message(
        characteristic.readValue(),
        expected,
        'readValue() should reject.')),
    test_desc);
</script>
